+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased changes
 
 - Introduce Entrypoint and Parameter types, and their owned versions.
+- Remove schema for state and add a schema for return values of init and receive functions.
 
 ## concordium-contracts-common 2.0.0 (2022-01-05)